DEV-UPDATE(merge_atm_mol) Merged the atomic and molecular versions of SFA and ADK `Samplers`. SFAAE is still waiting for further theoretical investigations.
