+++ src/ElectronSamplers/shared_methods.jl
